uitgebreid testen

+++ VrijwilligerApp/src/main/java/com/uhasselt/VrijwilligerApp/controllers/InschrijvingController.java
@@ -29,7 +29,7 @@ public class InschrijvingController {
         return null;
     }
 
-    public ResponseEntity<List<Inschrijving>> getAanwezigheden(int inschrijvingsId){
+    public ResponseEntity<List<Inschrijving>> getAanwezigheden(long inschrijvingsId){
         return null;
     }
