+++ VrijwilligerApp/src/main/java/com/uhasselt/VrijwilligerApp/interfaces/IInschrijvingService.java
@@ -17,8 +17,8 @@ public interface IInschrijvingService {
     Inschrijving verwijderBenodigheid(long benodigheidId, long inschrijvingId);
     Inschrijving schrijfInVoorEvenement(long groepID, long inschrijvingId);
     Inschrijving getInschrijving(int inschrijvingsId);
-    List<Inschrijving> getAanwezigheden(int vrijwilligersId);
+    List<Inschrijving> getAanwezigheden(long vrijwilligersId);
     boolean putAanwezigheid(long inschrijvingsId,boolean aanwezigheid);
-    boolean putScore(int vrijwilligersId, double score);
+    boolean putScore(long vrijwilligersId, double score);
 }
